refactor: 💡 worksSlideとreviewSlideの統合
