+++ src/pages/top/components/ReviewSlider/const/distanceToHero.ts
@@ -1,0 +1,1 @@
+export const REVIEW_DISTANCE_TO_HERO = 0;
