+++ src/pages/top/logics/generateSlideStyle/index.ts
@@ -9,7 +9,7 @@ export type BaseSlideList = {
 
 type SlideStyle = Pick<
   React.CSSProperties,
-  'width' | 'height' | 'left' | 'zIndex'
+  'width' | 'height' | 'right' | 'zIndex'
 >;
 
 type generateSlideStyleResult = {
@@ -19,7 +19,7 @@ type generateSlideStyleResult = {
  * 受け取った引数に対応するスタイルを返す関数
  *
  * @param {boolean} isHero - ヒーロースライドかどうかの判定
- * @param {number} leftIndex - スライドの左端インデックス
+ * @param {number} rightIndex - スライドの左端インデックス
  * @param {number} basicWidth - 基本スライドの幅
  * @param {number} heroWidth - ヒーロースライドの幅
  * @param {number} basicHeight - 基本スライドの高さ
@@ -27,12 +27,12 @@ type generateSlideStyleResult = {
  * @param {number} gap - スライド間の隙間
  * @param {number} DISTANCE_TO_HERO - ヒーロースライドまでの距離
  * @param {Array<Object>} slideList - スライドリスト（オブジェクトの配列）
- * @returns {{width: string, height: string, left: string, zIndex: string}} style オブジェクト
+ * @returns {{width: string, height: string, right: string, zIndex: string}} style オブジェクト
  */
 // todo: reviewスライド作成時に関数統合
 export const generateSlideStyle = (
   isHero: boolean,
-  leftIndex: number,
+  rightIndex: number,
   {
     basicWidth,
     heroWidth,
@@ -46,18 +46,18 @@ export const generateSlideStyle = (
   if (0 <= DISTANCE_TO_HERO && DISTANCE_TO_HERO > 3) {
     throw new Error('DISTANCE_TO_HEROがスライド枚数こえちょる');
   }
-  const isHeroOrLeft = leftIndex <= DISTANCE_TO_HERO;
+  const isHeroOrLeft = rightIndex <= DISTANCE_TO_HERO;
 
-  let left: number;
+  let right: number;
   let width = basicWidth;
   let height = basicHeight;
 
-  // leftを渡す
+  // rightを渡す
   if (isHeroOrLeft) {
-    left = leftIndex * (basicWidth + gap);
+    right = rightIndex * (basicWidth + gap);
   } else {
     const diff = heroWidth - basicWidth;
-    left = leftIndex * (basicWidth + gap) + diff;
+    right = rightIndex * (basicWidth + gap) + diff;
   }
 
   // widthとheightを渡す
@@ -66,15 +66,15 @@ export const generateSlideStyle = (
     height = heroHeight;
   }
 
-  const isFrontier = leftIndex + 1 === slideList.length;
+  const isFrontier = rightIndex + 1 === slideList.length;
 
-  // 代入されたleft,width,heightとz-indexをreturn
+  // 代入されたright,width,heightとz-indexをreturn
   const heroZIndex = 10;
   const frontier = 0;
   const style: SlideStyle = {
     width: `${width}px`,
     height: `${height}px`,
-    left: `${left}px`,
+    right: `${right}px`,
     zIndex: isFrontier ? frontier : heroZIndex,
   };
   return { style };
