+++ src/pages/top/logics/getRightIndex/fixtures.ts
